Step11&&Step12 김지민 - eCommerce

+++ src/main/java/h99/ecommerce/repository/PointRepository.java
@@ -8,7 +8,7 @@ public interface PointRepository {
 
     Point save(Point point);
 
-    Point findOne(int pointId);
+    Point findOne(Long pointId);
 
-    List<Point> findByUserId(int userId);
+    List<Point> findByUserId(Long userId);
 }
